restructure + rename files
