+++ src/components/field-input/generic-field-input.tsx
@@ -1,11 +1,11 @@
-import { IField } from "@/interfaces/form-interfaces";
+import { IField } from "@/interfaces/form";
 import DateFieldInput, { DateFieldInputProps } from "./date-field-input";
 import NumberFieldInput, { NumberFieldInputProps } from "./number-field-input";
 import TextFieldInput, { TextFieldInputProps } from "./text-field-input";
 import SelectFieldInput, { SelectFieldInputProps } from "./select-field-input";
 
 export type FieldInputProps = (DateFieldInputProps | NumberFieldInputProps | TextFieldInputProps | SelectFieldInputProps);
-export default function FieldInput(props: FieldInputProps) {
+export default function GenericFieldInput(props: FieldInputProps) {
   switch (props.type) {
     case "date":
       return (<DateFieldInput {...(props as DateFieldInputProps)}></DateFieldInput>);
