+++ src/components/field-input/text-field-input.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { IField } from "@/interfaces/form-interfaces";
+import { IField } from "@/interfaces/form";
 import { TextField, debounce } from "@mui/material";
 import { useState } from "react";
